support using a hash for docker::run $env param
Hashes are the much preferred way of handing env vars as the accurately represent what an environment is. An environment is a list of names and values associated with each name. A name cannot be duplicated.
Using an array allows duplication. You can have ['foo=bar','foo=baz'] which is a perfectly valid array, but not valid for environment variables.
Using a hash also allows one to merge 2 hashes together and get an expected result on collisions (latter wins).

In this change, arrays are still supported, we just add support for a hash as well.