stop a docker::run upstart resource when docker stops
Previously the upstart script generated by docker::run would only stop when the box is shutting down. This leads to issues as if docker shuts down, the container is killed without upstart expecting it. This results in upstart trying to restart the container when docker isn't running.

The solution is to shut down all containers before docker stops. On shutdown, docker itself will stop, and all the containers will stop because of that.