remove `hasrestart => true` from docker upstart service
Upstart has a long history of bugs and issues with restarting services. The issue at hand here is that when an upstart script is modified, and you restart the service, it doesn't pick up the modified script. Meaning if you change any of the parameters to the 'docker' puppet class that results in modifying the docker service, they won't take effect.

When you want to pick up changes to an upstart script you have to do a full stop & start.